fix logout button create session library for easy session management
